feat: add init file in init folder
